Add files via upload
Its a relatively large folder because it contains a plethora of images. The main focus of this homework was to do feature subset selection which I did through SFS and BFS methods and checked via cross validation. Other exercises involve linear embedding to produce a map of the world and the United States, K-means clustering to reconstruct an image of soccer players, and another question doing "Snapshot" PCA to analyze the eigenfaces of some of the faculty at Texas A&M.